requirements.txt
langchain-community
langchain-openai
pyyaml
tqdm
python-magic
python-dotenv
pytest
string2dict
pandas
